pyproject.toml
[project]
name = "robot-llm"
version = "0.1.0"
description = "机器人操作编排系统，集成 AI 自然语言任务规划"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.6.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "opencv-python>=4.13.0.92",
    "pyserial>=3.5",
    "ultralytics>=8.4.36",
    "scikit-learn>=1.8.0",
    "robotic-arm>=1.1.5",
]

[dependency-groups]
dev = []
